feat(cms): setup collections and listings for the dynamic views
fix #54

- add Pre-Publish Hooks

- artist name should be computed

- last name goes first in artist circles

- remove computed field from artist.json

- presave hook should only compute for artists

- seed data for testing

- Add listing layout json files

That way we don't have to set the layout for each category on multiple places. It'll make it easy to change things later.

- Turn off permalinks for news listing collections

We don't need this collection to generate pages. We'll use the most recent one on the main page.

- Add permalinks to testing md files

- amend artist listing (instead of artists)

- amend json files

- Remove content string from generated urls

This will allow us to keep the content folder to organize content, while outputting urls without the string

- ignore local todo notes

- Add listing pages for artfairs, artists, and exhibitions